run_avr: Add -v argument to raise verbosity level
Pass it more than once to raise it further. This works better with the
standard conventions, which would allow passing -vvv (instead of -v -v
-v). Unfortunately, this does not work here.

The type of avr->log has been altered to accomodate values up to 3.

+++ simavr/sim/sim_avr.h
@@ -245,7 +245,7 @@ typedef struct avr_t {
 	avr_int_table_t	interrupts;
 
 	// DEBUG ONLY -- value ignored if CONFIG_SIMAVR_TRACE = 0
-	int		trace : 1,
+	uint8_t	trace : 1,
 			log : 2; // log level, default to 1
 
 	// Only used if CONFIG_SIMAVR_TRACE is defined
